requirements.txt
mysql-connector-python==9.5.0
python-telegram-bot==22.5
python-dotenv==1.2.1

